exploring-multi2vec-clip-with-Python-and-flask/requirements.txt
Flask==2.0.2
weaviate_client==3.3.3
Werkzeug==2.0.1